requirements.txt
# ===========================================
# ALMIRA BACKEND - Python Dependencies
# ===========================================

# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Supabase
supabase==2.10.0

# Data Validation
pydantic==2.9.0
pydantic-settings==2.5.0
email-validator==2.2.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0

# HTTP Client
httpx==0.27.2

# Database (optional - uncomment if needed)
# asyncpg==0.29.0
# sqlalchemy==2.0.35

# Caching
redis==5.1.0
hiredis==2.3.2

# Rate Limiting
slowapi==0.1.9

# File Handling
# python-magic==0.4.27
pillow==10.4.0
# aiofiles==24.1.0

# Utilities
python-dotenv==1.0.1
orjson==3.10.7

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2

# Logging
loguru==0.7.2

# Development
black==24.8.0
isort==5.13.2
flake8==7.1.1
